move TestCpgBase outside
